# api/docs/swagger.yaml
definitions:
...
  models.CreateIncomeProduct:
    properties:
      count:
        type: integer
      income_id:
        type: string
      price:
        type: number
      product_id:
        type: string
    type: object
  models.CreateProduct:
    properties:
      bar_code:
        type: string
      category_id:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
...
  models.IncomeProduct:
    properties:
      count:
        type: integer
      created_at:
        type: string
      id:
        type: string
      income_id:
        type: string
      price:
        type: number
      product_id:
        type: string
      updated_at:
        type: string
    type: object
  models.IncomeProductsResponse:
    properties:
      count:
        type: integer
      incomeProducts:
        items:
          $ref: '#/definitions/models.IncomeProduct'
        type: array
    type: object
...
  models.Product:
    properties:
      bar_code:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      updated_at:
        type: string
    type: object
  models.ProductsResponse:
    properties:
      count:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      status_code:
        type: integer
    type: object
...
  models.UpdateIncomeProduct:
    properties:
      count:
        type: integer
      id:
        type: string
      income_id:
        type: string
      price:
        type: number
      product_id:
        type: string
    type: object
  models.UpdateProduct:
    properties:
      category_id:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /income_product:
    post:
      consumes:
      - application/json
      description: create a new income product
      parameters:
      - description: income_product
        in: body
        name: income_product
        schema:
          $ref: '#/definitions/models.CreateIncomeProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IncomeProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Creates a new income product
      tags:
      - income_product
  /income_product/{id}:
    delete:
      consumes:
      - application/json
      description: delete income_product
      parameters:
      - description: income_product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete income_product
      tags:
      - income_product
    get:
      consumes:
      - application/json
      description: get income product by id
      parameters:
      - description: income_product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IncomeProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get income product by id
      tags:
      - income_product
    put:
      consumes:
      - application/json
      description: update income product
      parameters:
      - description: income_product_id
        in: path
        name: id
        required: true
        type: string
      - description: income_product
        in: body
        name: income_product
        schema:
          $ref: '#/definitions/models.UpdateIncomeProduct'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IncomeProduct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update income product
      tags:
      - income_product
  /income_products:
    get:
      consumes:
      - application/json
      description: get income product list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IncomeProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get income product list
      tags:
      - income_product
...
  /product:
    post:
      consumes:
      - application/json
      description: create a new product
      parameters:
      - description: product
        in: body
        name: product
        schema:
          $ref: '#/definitions/models.CreateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new product
      tags:
      - product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: delete product
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete product
      tags:
      - product
    get:
      consumes:
      - application/json
      description: get product by id
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get product by id
      tags:
      - product
    put:
      consumes:
      - application/json
      description: update product
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      - description: product
        in: body
        name: product
        schema:
          $ref: '#/definitions/models.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update product
      tags:
      - product
  /products:
    get:
      consumes:
      - application/json
      description: get product list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get product list
      tags:
      - product
...
swagger: "2.0"